mod_tls: avoid infinite loop in tls_core cleanup
In `mod_tls''s `tls_core.c` if there is an error in
`setup_hello_config()` control flow is redirected to the `cleanup`
label to log the error. However, the body of that logic includes its own
`goto cleanup` at the end. Since the `rustls_result` guarding this logic
won't have changed since the original error this will produce an
infinite loop, forever jumping back to the start of the `cleanup`
section.